Gather the MuJoCo parity tests of equality constraints and friction loss under the parity module.
